Refactor library
This commit refactors the library to be more managable and readable to humans.
This should improve the development ease of use.
